Big reorganization of internal structures in the compiz plugin.
The lcms profiles are created and cached much more then before. Every
time the window regions are updated, the compiz plugin looks up the
associated profiles instead of creating the lcms profile everytime
the client sends a _NET_COLOR_MANAGEMENT message.
Also, profiles are now refcounted and can be deleted from the plugin.
This is done by sending a XcolorProfile with a zero-length to the root
window. It's currently untested though.