client: Change target output at keypress.
This needed a small change in the compiz plugin as well.

To test this, upload different profiles into randr outputs, then start
the client and press 'space' once to activate the color management.
Now you can press 'delete' to switch the target outputs and hopefully
you'll see that the region slightly changes color.